extract serialization logic from framework to serialization.rs
